requirements.txt
langchain==0.1.9
openai==1.12.0
litellm==1.27.6
langchain_openai==0.0.7
langchain-experimental==0.0.52
fastapi==0.109.2
uvicorn==0.27.1
atlassian-python-api==3.41.10
qdrant-client==1.7.3
sentence-transformers==2.4.0
tokenizers==0.15.2
unstructured[all-docs]==0.12.4
websockets==12.0
urllib3==2.2.1
tiktoken==0.6.0
requests==2.31.0
PyPDF2==3.0.1
chainlit==1.0.301
aioconsole==0.7.0
youtube-transcript-api==0.6.1
google-api-python-client==2.109.0